MVC changed to MVVM, codestyle refactoring, urlSession configuration added
