+++ RickAndMortyApp/Model/Client.swift
@@ -11,6 +11,8 @@ import Foundation
 
 struct Client {
     
+    // MARK: - Public
+    
     func character() -> Character {
         let character = Character(client: self)
         return character
